# schemas/openapi.json
{
  "components": {
    "schemas": {
      "AssetInfo": {
        "additionalProperties": false,
        "description": "`POST /api/assets` 의 응답. content-addressed 이므로 같은 파일은 같은 해시다.",
        "properties": {
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "hash": {
            "description": "내용 해시. `GET /api/assets/{hash}` 의 키",
            "title": "Hash",
            "type": "string"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "픽셀 높이. 이미지가 아니면 null",
            "title": "Height"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "픽셀 너비. 이미지가 아니면 null",
            "title": "Width"
          }
        },
        "required": [
          "hash",
          "size_bytes",
          "media_type"
        ],
        "title": "AssetInfo",
        "type": "object"
      },
...
      "Body_graph_from_png_api_graph_from_png_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "`nodal_workflow` iTXt 청크를 담은 PNG",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_graph_from_png_api_graph_from_png_post",
        "type": "object"
      },
      "Body_upload_asset_api_assets_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "업로드할 파일",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_asset_api_assets_post",
        "type": "object"
      },
...
      "CreateRunResponse": {
        "additionalProperties": false,
        "description": "`POST /api/runs` — 큐에 넣고 즉시 돌아온다. 진행은 `/ws` 로 본다.",
        "properties": {
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "CreateRunResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "모든 4xx·5xx 응답의 본문. 상태 코드로 분기하고 본문은 항상 이 모양이다.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GraphFromPngResponse": {
        "additionalProperties": false,
        "description": "`POST /api/graph/from-png` 의 응답 (M3).\n\nPNG 의 `nodal_workflow` iTXt 청크에서 캐논 그래프를 꺼낸 결과다.\n파싱을 서버 한 곳에만 두는 이유는 iTXt 파서가 Python·TS 양쪽에 생기면\n그것이 곧 \"규칙을 두 번 쓰지 않는다\" 위반이기 때문이다.\n\n청크가 없거나 JSON 이 깨졌으면 이 응답이 아니라 `ErrorResponse` 가 나간다.",
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/Graph",
            "description": "복원된 캐논 그래프"
          },
          "nodal_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "PNG 에 함께 박힌 `nodal_version` iTXt. 없으면 null",
            "title": "Nodal Version"
          }
        },
        "required": [
          "graph"
        ],
        "title": "GraphFromPngResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidateRequest": {
        "additionalProperties": false,
        "properties": {
          "graph": {
            "$ref": "#/components/schemas/Graph",
            "description": "캐논 그래프 문서 그대로. 서버는 변환하지 않는다"
          },
          "outputs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "검증할 출력 노드. 생략하면 그래프의 `outputs`",
            "title": "Outputs"
          }
        },
        "required": [
          "graph"
        ],
        "title": "ValidateRequest",
        "type": "object"
      },
      "ValidateResponse": {
        "additionalProperties": false,
        "description": "`POST /api/graph/validate` — 실행 없이 검증만 한다.\n\n무효한 그래프도 **200** 이다. 검증은 질의이지 명령이 아니므로, \"이 그래프는\n무효다\"는 성공적인 답이다. 4xx 는 요청 자체가 잘못됐을 때만 나간다.",
        "properties": {
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueModel"
            },
            "title": "Issues",
            "type": "array"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid"
        ],
        "title": "ValidateResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "캐논 그래프를 그대로 주고받는다. 변환 레이어가 없다.\n\n모든 실패 응답의 본문은 `ErrorResponse` 이고, `issues[].location` 이 `nodes.<id>.inputs.<socket>` 경로를 준다 — 프론트는 이것으로 캔버스의 해당 소켓을 지목한다.",
    "summary": "노드 그래프 실행 서버",
    "title": "nodal API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/assets": {
      "post": {
        "description": "content-addressed 로 저장한다. 같은 내용이면 같은 해시가 돌아온다.",
        "operationId": "upload_asset_api_assets_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_asset_api_assets_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "요청이 잘못됐다"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "대상이 없다"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "그래프 검증 실패. `issues` 가 어느 노드·어느 소켓인지 지목한다"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "계약은 확정됐지만 아직 구현되지 않았다 (M3)"
          }
        },
        "summary": "입력 파일 업로드",
        "tags": [
          "assets"
        ]
      }
    },
    "/api/assets/{asset_hash}": {
      "get": {
        "operationId": "get_asset_api_assets__asset_hash__get",
        "parameters": [
          {
            "description": "내용 해시",
            "in": "path",
            "name": "asset_hash",
            "required": true,
            "schema": {
              "description": "내용 해시",
              "title": "Asset Hash",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "파일 내용 그대로. `Content-Type` 은 업로드 시 판별된 것"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "요청이 잘못됐다"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "대상이 없다"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "그래프 검증 실패. `issues` 가 어느 노드·어느 소켓인지 지목한다"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "계약은 확정됐지만 아직 구현되지 않았다 (M3)"
          }
        },
        "summary": "content-addressed 조회",
        "tags": [
          "assets"
        ]
      }
    },
...
    "/api/graph/from-png": {
      "post": {
        "description": "PNG 의 `nodal_workflow` iTXt 청크에서 캐논 그래프를 꺼낸다. 프론트의 드래그앤드롭이 이 엔드포인트로 파일을 던진다 — iTXt 파서를 Python·TS 양쪽에 두지 않기 위해서다.\n\n**M3 계약 시점에는 아직 구현되지 않았다.** 501 을 돌려준다.",
        "operationId": "graph_from_png_api_graph_from_png_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_graph_from_png_api_graph_from_png_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphFromPngResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "요청이 잘못됐다"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "대상이 없다"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "그래프 검증 실패. `issues` 가 어느 노드·어느 소켓인지 지목한다"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "계약은 확정됐지만 아직 구현되지 않았다 (M3)"
          }
        },
        "summary": "PNG 에서 워크플로 복원",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/graph/validate": {
      "post": {
        "description": "구조 · 노드 타입 · 소켓 타입 호환성을 검사한다. **무효한 그래프도 200** 이다 — 검증은 질의이지 명령이 아니므로 `valid: false` 가 성공적인 답이다.",
        "operationId": "validate_graph_endpoint_api_graph_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "요청이 잘못됐다"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "대상이 없다"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "그래프 검증 실패. `issues` 가 어느 노드·어느 소켓인지 지목한다"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "계약은 확정됐지만 아직 구현되지 않았다 (M3)"
          }
        },
        "summary": "실행 없이 검증만",
        "tags": [
          "graph"
        ]
      }
    },
...
    "/api/runs": {
      "get": {
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "description": "히스토리 최대 개수",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "히스토리 최대 개수",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "큐와 히스토리",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "큐에 넣고 즉시 돌아온다. 진행 상황은 `/ws` 로 본다.\n\n큐 진입 전에 전체 검증을 통과해야 한다 — 실패하면 422 와 `issues`.",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "요청이 잘못됐다"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "대상이 없다"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "그래프 검증 실패. `issues` 가 어느 노드·어느 소켓인지 지목한다"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "계약은 확정됐지만 아직 구현되지 않았다 (M3)"
          }
        },
        "summary": "실행 큐에 등록",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}": {
      "delete": {
        "description": "협조적 취소를 요청한다. 이미 끝난 실행이면 그때의 상태가 그대로 돌아온다.",
        "operationId": "cancel_run_api_runs__run_id__delete",
        "parameters": [
          {
            "description": "실행 ID",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "실행 ID",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "요청이 잘못됐다"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "대상이 없다"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "그래프 검증 실패. `issues` 가 어느 노드·어느 소켓인지 지목한다"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "계약은 확정됐지만 아직 구현되지 않았다 (M3)"
          }
        },
        "summary": "취소",
        "tags": [
          "runs"
        ]
      },
      "get": {
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "description": "실행 ID",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "실행 ID",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "요청이 잘못됐다"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "대상이 없다"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "그래프 검증 실패. `issues` 가 어느 노드·어느 소켓인지 지목한다"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "계약은 확정됐지만 아직 구현되지 않았다 (M3)"
          }
        },
        "summary": "상태와 결과",
        "tags": [
          "runs"
        ]
      }
    }
  },
...
}